+++ R/precision.R
@@ -8,6 +8,9 @@
 #'
 #' @description
 #' 
+#' The goal of \code{precision} is to allow you to study the relationship between shrinkage of ridge
+#' regression coefficients and their precision directly by calculating measures of each.
+#' 
 #' Three measures of (inverse) precision based on the \dQuote{size} of the
 #' covariance matrix of the parameters are calculated. Let \eqn{V_k \equiv \text{Var}(\mathbf{\beta}_k)} 
 #' be the covariance matrix for a given ridge constant, and let \eqn{\lambda_i , i= 1,
@@ -20,6 +23,16 @@
 #'   \item \code{"trace"}: \eqn{ \text{trace}( V_k ) = \sum \lambda} corresponds conceptually to Pillai's trace criterion 
 #'   \item \code{"max.eig"}: \eqn{ \lambda_1 = \max (\lambda)} corresponds to Roy's largest root criterion.  
 #' }
+#' 
+#' Two measures of shrinkage are also calculated:
+#' \itemize{
+#'    \item \code{norm.beta}: the root mean square of the coefficient vector \eqn{\lVert\mathbf{\beta}_k \rVert},
+#'         normalized to a maximum of 1.0 if \code{normalize == TRUE} (the default).
+#'    \item \code{norm.diff} the root mean square of the difference 
+#'         \eqn{\lVert \mathbf{\beta}_0 - \mathbf{\beta}_k \rVert}. This measure is inversely related to \code{norm.beta}
+#' }
+#' 
+#' A plot method, \code{link{plot.precision}} facilitates making graphs of these quantities.
 #' 
 #' @param object An object of class \code{ridge} or \code{lm}
 #' @param det.fun Function to be applied to the determinants of the covariance
@@ -35,6 +48,8 @@
 #' \item{trace}{The trace of the covariance matrix}
 #' \item{max.eig}{Maximum eigen value of the covariance matrix}
 #' \item{norm.beta}{The root mean square of the estimated coefficients, possibly normalized} 
+#' \item{norm.diff}{The root mean square of the difference between the OLS solution 
+#'       (\code{lambda = 0}) and ridge solutions} 
 #' 
 #' @note Models fit by \code{lm} and \code{ridge} use a different scaling for
 #' the predictors, so the results of \code{precision} for an \code{lm} model
@@ -109,14 +124,14 @@ precision.ridge <- function(object,
 	norm <- sqrt(rowMeans(coef(object)^2))
 	if (normalize) norm <- norm / max(norm)
 	
-	# calculate bias: norm of (beta[lambda=0] - beta)
+	# calculate norm.diff: norm of (beta[lambda=0] - beta)
 	b0index <- which(object$lambda == 0)
 	if (b0index != 0) {
 	  b0 <- b[b0index, ]
 	  dif <- sweep(b, 2, b0)
-	  bias <- sqrt(rowMeans(dif^2))
+	  norm.diff <- sqrt(rowMeans(dif^2))
 	}
-	else warning("There is no OLS solution (lambda==0), so can't calculate bias")
+	else warning("There is no OLS solution (lambda==0), so can't calculate norm.diff")
 
 	res <- data.frame(lambda=object$lambda, 
 	           df=object$df, 
@@ -124,7 +139,7 @@ precision.ridge <- function(object,
 	           trace=trace, 
 	           max.eig=meig, 
 	           norm.beta=norm,
-	           bias=bias)
+	           norm.diff=norm.diff)
 	class(res) <- c("precision", "data.frame")
 	res
 }