+++ NEWS.md
@@ -4,7 +4,8 @@ o Fleshed out and installed `plot.precision()`
 o Illustrate `plot.precision()` in `README.Rmd`
 o Added `diab` data, diabetes from CASI
 o Fixed documentation for `plot.ridge()` so that `plot.pcaridge()` is documented explicitly.
-o 'vif.ridge()` now returns a data.frame of class also "vif.ridge" to prepare for a plot method.
+o 'vif.ridge()` now returns a "vif.ridge" object list to prepare for a plot method.
+o Added `plot.vif.ridge()` to plot VIFs vs shrinkage
 
 ## genridge 0.7.1 (2024-11-07)